Move to trainee-manager/ folder
